requirements.txt
discord.py
flask
python-dotenv
schedule
requests
pandas
